+++ app/exception_models.py
@@ -1,0 +1,8 @@
+class NoExistingUser(Exception):
+    def __init__(self, message):
+        Exception.__init__(self, message)
+
+
+class InvalidUserInfo(Exception):
+    def __init__(self, message):
+        Exception.__init__(self, message)
