+++ tests/test_app.py
@@ -3,7 +3,7 @@ import json
 
 import pytest
 
-from app import create_app, db, models
+from app import create_app, db, user_model, user_manager_model, community_resource_model, exception_models
 
 
 @pytest.fixture
@@ -49,7 +49,7 @@ def test_get_greeting(client):
     username = "foo"
     password = "bar"
 
-    user = models.User(username=username)
+    user = user_model.User(username=username)
     user.password = password
 
     db.session.add(user)
@@ -69,7 +69,7 @@ def test_post_greeting(client):
     username = "foo"
     password = "bar"
 
-    user = models.User(username=username)
+    user = user_model.User(username=username)
     user.password = password
 
     db.session.add(user)
@@ -90,7 +90,7 @@ def test_get_user(client):
     username = "foo"
     password = "bar"
 
-    user = models.User(username=username)
+    user = user_model.User(username=username)
     user.password = password
 
     db.session.add(user)
@@ -127,7 +127,7 @@ def test_get_user_token(client):
     username = "foo"
     password = "bar"
 
-    user = models.User(username=username)
+    user = user_model.User(username=username)
     user.password = password
 
     db.session.add(user)
@@ -140,7 +140,7 @@ def test_get_user_token(client):
     body = json.loads(rv.get_data(as_text=True))
 
     assert rv.status_code == 201
-    assert models.User.verify_auth_token(
+    assert user_model.User.verify_auth_token(
         body["token"]).username == user.username
 
 
@@ -155,7 +155,7 @@ def test_post_user(client):
     }))
 
     assert rv.status_code == 201
-    assert models.User.get_user_by_username(username) is not None
+    assert user_model.User.get_user_by_username(username) is not None
 
     # username must be unique
     rv = client.post("/user", headers=get_headers(), data=json.dumps({
