+++ tests/test_models.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from app import models
+from app import user_model, user_manager_model, community_resource_model, exception_models
 
 
 @patch("flask_sqlalchemy.SignallingSession", autospec=True)
@@ -10,7 +10,7 @@ def test_user_password(mock_session):
     username = "foo"
     password = "bar"
 
-    user = models.User(id=1, username=username)
+    user = user_model.User(id=1, username=username)
 
     user.password = password
 
@@ -25,13 +25,13 @@ def test_user_password(mock_session):
         print(user.password)
 
 
-@patch("app.models.db.Model.query")
-@patch("app.models.current_app")
+@patch("app.user_model.db.Model.query")
+@patch("app.user_model.current_app")
 @patch("flask_sqlalchemy.SignallingSession", autospec=True)
 def test_user_token(mock_session, mock_current_app, mock_query):
     username = "foo"
     password = "bar"
-    user = models.User(id=1, username=username)
+    user = user_model.User(id=1, username=username)
     user.password = password
 
     mock_current_app.config = {"SECRET_KEY": "secret-key"}
@@ -39,24 +39,24 @@ def test_user_token(mock_session, mock_current_app, mock_query):
 
     token = user.generate_auth_token()
 
-    user_ = models.User.verify_auth_token(token)
+    user_ = user_model.User.verify_auth_token(token)
 
     # correct token
     assert user == user_
 
-    user_ = models.User.verify_auth_token("afdlkjsls;kfd")
+    user_ = user_model.User.verify_auth_token("afdlkjsls;kfd")
 
     # incorrect token
     assert user_ is None
 
     token = user.generate_auth_token(expiration=-1)
-    user_ = models.User.verify_auth_token(token)
+    user_ = user_model.User.verify_auth_token(token)
 
     # expired token
     assert user_ is None
 
     mock_current_app.config = {"SECRET_KEY": "secret-key2"}
-    user_ = models.User.verify_auth_token(token)
+    user_ = user_model.User.verify_auth_token(token)
 
     # different key
     assert user_ is None
